# backend/swagger.json
{"openapi":"3.0.0","info":{"title":"My API","version":"1.0.0","description":"Documentation using Swagger"},"servers":[{"url":"http://localhost:3000","description":"Local server"}],"paths":{"/users":{"get":{"summary":"Retrieves all users","tags":["Users"],"responses":{"200":{"description":"Successfully retrieved the list of users.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","example":1},"name":{"type":"string","example":"John Doe"}}}}}}}}},"post":{"summary":"Creates a new user","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Maria Johnson"}}}}}},"responses":{"201":{"description":"User successfully created.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":2},"name":{"type":"string","example":"Maria Johnson"}}}}}}}}},"/users/{id}":{"get":{"summary":"Retrieves a single user by ID","tags":["Users"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"User ID"}],"responses":{"200":{"description":"User successfully retrieved.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer","example":1},"name":{"type":"string","example":"John Doe"}}}}}},"404":{"description":"User not found"}}},"put":{"summary":"Updates a user by ID","tags":["Users"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"User ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Anna Williams"}}}}}},"responses":{"200":{"description":"User successfully updated."}}},"delete":{"summary":"Deletes a user by ID","tags":["Users"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"User ID"}],"responses":{"204":{"description":"User successfully deleted."}}}}},"components":{},"tags":[{"name":"Users","description":"User management"}]}